+++ UIAPlatform.core/src/main/java/module-info.java
@@ -1,6 +1,0 @@
-module at.bestsolution.uia.core {
-    requires transitive javafx.graphics;
-
-    exports at.bestsolution.uia.core;
-//    opens at.bestsolution.uia.core;
-}