+++ UIAPlatform/src/main/java/at/bestsolution/uia/internal/PlatformBoundsUtil.java
@@ -8,7 +8,6 @@ import com.sun.glass.ui.Screen;
 
 import javafx.geometry.Point2D;
 import javafx.scene.Scene;
-import javafx.stage.Window;
 
 
 public class PlatformBoundsUtil {
@@ -16,18 +15,18 @@ public class PlatformBoundsUtil {
   private static final Logger LOG = Logger.create(PlatformBoundsUtil.class);
 
 
-  public static float getPlatformScaleX(Window window) {
-    return WindowHelper.getWindowAccessor().getPlatformScaleX(window);
-  }
-  public static float getPlatformScaleX(Scene scene) {
-    return getPlatformScaleX(scene.getWindow());
-  }
-  public static float getPlatformScaleY(Window window) {
-    return WindowHelper.getWindowAccessor().getPlatformScaleY(window);
-  }
-  public static float getPlatformScaleY(Scene scene) {
-    return getPlatformScaleY(scene.getWindow());
-  }
+  // public static float getPlatformScaleX(Window window) {
+  //   return WindowHelper.getWindowAccessor().getPlatformScaleX(window);
+  // }
+  // public static float getPlatformScaleX(Scene scene) {
+  //   return getPlatformScaleX(scene.getWindow());
+  // }
+  // public static float getPlatformScaleY(Window window) {
+  //   return WindowHelper.getWindowAccessor().getPlatformScaleY(window);
+  // }
+  // public static float getPlatformScaleY(Scene scene) {
+  //   return getPlatformScaleY(scene.getWindow());
+  // }
 
   public static Point2D convertToPlatformPoint(Scene scene, Point2D p) {
     Screen screen = getScreen(scene);
@@ -40,7 +39,7 @@ public class PlatformBoundsUtil {
     return new Point2D(screen.fromPlatformX((int) Math.round(p.getX())), screen.fromPlatformY((int) Math.round(p.getY())));
   }
 
-  public static Screen getScreen(Scene scene) {
+  private static Screen getScreen(Scene scene) {
     if (scene == null || scene.getWindow() == null) return null;
     TKStage tkStage = WindowHelper.getPeer(scene.getWindow());
     if (!(tkStage instanceof WindowStage)) return null;